+++ src/OrderScrapper/OrderScrapper/Core/BSWebNavigationDelegate.swift
@@ -37,18 +37,18 @@ internal class BSWebNavigationDelegate: NSObject, WKNavigationDelegate {
     }
     
     internal func webView(_ webView: WKWebView, didFail navigation: WKNavigation!, withError error: Error) {
-        SentrySDK.capture(error: error)
+        FirebaseAnalyticsUtil.logSentryError(error: error)
         self.didFailNavigation(for: webView.url, withError: error)
     }
     
     internal func webView(_ webView: WKWebView, didFailProvisionalNavigation navigation: WKNavigation!, withError error: Error) {
-        SentrySDK.capture(error: error)
+        FirebaseAnalyticsUtil.logSentryError(error: error)
         self.didFailNavigation(for: webView.url, withError: error)
     }
     
     private func didFailNavigation(for url: URL?, withError error: Error) {
         if let observer = self.observer {
-            SentrySDK.capture(error: error)
+            FirebaseAnalyticsUtil.logSentryError(error: error)
             observer.didFailPageNavigation(for: url, withError: error)
         }
     }