+++ src/OrderScrapper/OrderScrapper/Network/Response/ScrapperConfigModels/ScrapeConfigs.swift
@@ -1,0 +1,10 @@
+//  ScrapeConfigs.swift
+//  OrderScrapper
+
+
+import Foundation
+
+class ScrapeConfigs: Decodable {
+    let configurations: [PlatformSourceConfig]
+    let sentry: SentryConfigs
+}