+++ src/OrderScrapper/OrderScrapper/Network/Services/AmazonService.swift
@@ -36,7 +36,7 @@ class AmazonService {
                 if response.isError {
                     completionHandler(nil, APIError(error: response.error ?? "Error"))
                     print(AppConstants.tag, "getDateRange", response.error ?? "Error")
-                    SentrySDK.capture(error: APIError(error: response.error ?? Strings.ErrorAPIReponseDateRange))
+                    FirebaseAnalyticsUtil.logSentryError(error: APIError(error: response.error ?? Strings.ErrorAPIReponseDateRange))
                 } else {
                     completionHandler(response.data, nil)
                 }
@@ -67,7 +67,7 @@ class AmazonService {
                 if response.isError {
                     completionHandler(nil, APIError(error: response.error ?? "Error"))
                     print(AppConstants.tag, "uploadFile", response.error ?? "Error")
-                    SentrySDK.capture(error: APIError(error: response.error ?? Strings.ErrorAPIReposneUplodFile))
+                    FirebaseAnalyticsUtil.logSentryError(error: APIError(error: response.error ?? Strings.ErrorAPIReposneUplodFile))
                 } else {
                     completionHandler(response.data, nil)
                 }
@@ -87,7 +87,7 @@ class AmazonService {
                 if response.isError {
                     completionHandler(nil, APIError(error: response.error ?? "Error"))
                     print(AppConstants.tag, "getPIIList", response.error ?? "Error")
-                    SentrySDK.capture(error: APIError(error: response.error ?? Strings.ErrorAPIReposnePIIList))
+                    FirebaseAnalyticsUtil.logSentryError(error: APIError(error: response.error ?? Strings.ErrorAPIReposnePIIList))
                 } else {
                     completionHandler(response.data, nil)
                 }
@@ -108,7 +108,7 @@ class AmazonService {
                 if response.isError {
                     completionHandler(nil, APIError(error: response.error ?? "Error"))
                     print(AppConstants.tag, "getAccounts", response.error ?? "Error")
-                    SentrySDK.capture(error: APIError(error: response.error ?? Strings.ErrorAPIReposneGetAccount))
+                    FirebaseAnalyticsUtil.logSentryError(error:  APIError(error: response.error ?? Strings.ErrorAPIReposneGetAccount))
                 } else {
                     completionHandler(response.data, nil)
                 }
@@ -132,7 +132,7 @@ class AmazonService {
                 if response.isError {
                     completionHandler(nil, APIError(error: response.error ?? "Error"))
                     print(AppConstants.tag, "registerConnection", response.error ?? "Error")
-                    SentrySDK.capture(error: APIError(error: response.error ?? Strings.ErrorAPIReposneRegisterConnection))
+                    FirebaseAnalyticsUtil.logSentryError(error:  APIError(error: response.error ?? Strings.ErrorAPIReposneRegisterConnection))
                 } else {
                     completionHandler(response.data, nil)
                 }
@@ -157,7 +157,7 @@ class AmazonService {
                 if response.isError {
                     completionHandler(nil, APIError(error: response.error ?? "Error"))
                     print(AppConstants.tag, "updateStatus", response.error ?? "Error")
-                    SentrySDK.capture(error: APIError(error: response.error ?? Strings.ErrorAPIResponseUpdateStatus))
+                    FirebaseAnalyticsUtil.logSentryError(error: APIError(error: response.error ?? Strings.ErrorAPIResponseUpdateStatus))
                 } else {
                     completionHandler(response.data, nil)
                 }
@@ -168,13 +168,13 @@ class AmazonService {
         return client
     }
     
-   static func getScrapperConfig(orderSource: [String], completionHandler: @escaping ([PlatformSourceConfig]?, Error?) -> Void) -> APIClient {
+   static func getScrapperConfig(orderSource: [String], completionHandler: @escaping (ScrapeConfigs?, Error?) -> Void) -> APIClient {
         
-        let client = NetworkClient<APIResponse<[PlatformSourceConfig]>>(relativeURL: ScrapperConfigURL, requestMethod: .post)
+        let client = NetworkClient<APIResponse<ScrapeConfigs>>(relativeURL: ScrapperConfigURL, requestMethod: .post)
         client.body = [JSONKeys.configDetails.rawValue: orderSource]
         
         client.executeAPI() { (response, error) in
-            if let response = response as? APIResponse<[PlatformSourceConfig]> {
+            if let response = response as? APIResponse<ScrapeConfigs> {
                 if response.isError {
                     completionHandler(nil, APIError(error: response.error ?? "Error"))
                     print(AppConstants.tag, "getScrapperConfig", response.error ?? "Error")